// src/services/api/generated/models/system-configuration.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Test Management API
 * API for managing boards, tests, and execution.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface SystemConfiguration {
    'id': number;
    'key': string;
    'value'?: any;
    'description'?: string;
    'updated_at': string;
}